HOME-275 - Split admin space from dashboard
