+++ client/app.scss
@@ -7,3 +7,4 @@ $gc-icon-font-url: '~graphen/src/assets/font';
 @import "modules/AgentsStatus/styles";
 @import "modules/Dashboard/styles";
 @import "modules/AlertPanel/styles";
+@import "modules/Admin/styles";
