+++ client/modules/Admin/Alerts/index.jsx
@@ -2,22 +2,16 @@
 import { connect } from 'react-redux';
 import * as agentActions from 'client/models/agents/actions';
 import * as agentSelectors from 'client/models/agents/selectors';
-import ControlPanel from './ControlPanel';
+import Alerts from './Alerts';
 
 const mapStateToProps = state => ({
   isAlerts: agentSelectors.isAlerts(state),
 });
 
 const mapDispatchToProps = dispatch => ({
-  sniffAgents: () => {
-    dispatch(agentActions.sniffAgents());
-  },
   onToggle: () => {
     dispatch(agentActions.toggleAlerts());
   },
-  sendAlert: () => {
-    dispatch(agentActions.sendAlert());
-  },
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(ControlPanel);
+export default connect(mapStateToProps, mapDispatchToProps)(Alerts);
