+++ client/modules/Admin/Sniffer/Sniffer.jsx
@@ -1,0 +1,29 @@
+// @flow
+import React from 'react';
+import Switch from 'client/components/Switch';
+
+type Props = {
+  sniffAgents: () => void,
+};
+
+const Sniffer = (props: Props) => {
+  const { sniffAgents } = props;
+
+  return (
+    <div className="gc-card gc-card--default">
+      <div className="gc-panel">
+        <div className="gc-panel__title">Agent lookup</div>
+        <div className="gc-panel__content">
+          Agent lookup searches for compatible devices in your network.
+          <p>
+            <button className="gc-btn gc-btn--success" onClick={sniffAgents}>
+              Search for agents
+            </button>
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Sniffer;
