+++ client/modules/Admin/Sniffer/index.jsx
@@ -1,0 +1,14 @@
+// @flow
+import { connect } from 'react-redux';
+import * as agentActions from 'client/models/agents/actions';
+import Sniffer from './Sniffer';
+
+const mapStateToProps = state => ({});
+
+const mapDispatchToProps = dispatch => ({
+  sniffAgents: () => {
+    dispatch(agentActions.sniffAgents());
+  },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(Sniffer);
