--- client/modules/AgentsStatus/ControlPanel/_styles.scss
@@ -1,19 +1,0 @@
-.control-panel {
-  &__control {
-    @extend .c-control;
-  }
-
-  &__alerts {
-    @extend .gc-switch--success;
-
-    margin-left: 15px;
-    vertical-align: middle;
-  }
-
-  &__send-alert {
-    @extend .gc-btn;
-    @extend .gc-btn--success;
-
-    margin-left: 15px;
-  }
-}
