+++ client/modules/AgentsStatus/_styles.scss
@@ -27,4 +27,3 @@ $color-persimmon: #ff7157;
 }
 
 @import "./List/styles";
-@import "./ControlPanel/styles";
