+++ controllers/admin.go
@@ -1,0 +1,14 @@
+package controllers
+
+import (
+	"github.com/coda-it/gowebserver/router"
+	"github.com/coda-it/gowebserver/session"
+	"github.com/coda-it/gowebserver/store"
+	"github.com/smart-evolution/shpanel/utils"
+	"net/http"
+)
+
+// CtrAdmin - controller for agents list
+func CtrAdmin(w http.ResponseWriter, r *http.Request, opt router.UrlOptions, sm session.ISessionManager, s store.IStore) {
+	utils.RenderTemplate(w, r, "admin", sm, make(map[string]interface{}))
+}
