+++ main.go
@@ -39,6 +39,7 @@ func New(port string, p *persistence.Persistance) *WebServer {
 	}
 
 	server := gowebserver.New(serverOptions, controllers.NotFound)
+	server.Router.AddRoute("/admin", controllers.CtrAdmin)
 	server.Router.AddRoute("/agent/{agent}", controllers.CtrDashboard)
 	server.Router.AddRoute("/agent/{agent}/edit", controllers.CtrAgentEdit)
 	server.Router.AddRoute("/", controllers.CtrDashboard)
