+++ views/admin.html
@@ -1,0 +1,9 @@
+{{define "body"}}
+    <div class="gc-page">
+        <section>
+            <article class="js-app">
+                <div class="gc-loader"></div>
+            </article>
+        </section>
+    </div>
+{{end}}
